feature: move app creation to __init__.py
It is recommended by official Flask tutorial.
Using create_app() factory method makes it easier to test the app
and make it more complex later.

To run the app, issue in bash (from root directory of project):

export FLASK_APP=flask_app
export FLASK_ENV=development
flask run
